+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 ## 1.0.0
 
 - Initial version.
+
+## 1.1.0
+
+- Fixed issues with setting log level
+- Fixed issues with log level formatting
+- added formatter function
+- added tests
+- improved documentation